pyproject.toml
[tool.poetry]
name = "django-backup-restore"
version = "0.1.0"
description = "A generic, pluggable backup and restore command framework for Django."
authors = ["Brian Kohan <bckohan@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/bckohan/django-backup-restore"
homepage = "https://django-typer.readthedocs.io"
keywords = ["django", "CLI", "management", "backup", "restore", "commands"]
packages = [
    { include = "django" }
]

classifiers = [
    "Environment :: Console",
    "Framework :: Django",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules"
]


[tool.poetry.urls]
"Homepage" = "https://django-backup-restore.readthedocs.io"
"Documentation" = "https://django-backup-restore.readthedocs.io"
"Repository" = "https://github.com/bckohan/django-backup-restore"
"Issues" = "https://github.com/bckohan/django-backup-restore/issues"
"Changelog" = "https://django-backup-restore.readthedocs.io/en/latest/changelog.html"
"Code_of_Conduct" = "https://github.com/django-commons/membership/blob/main/CODE_OF_CONDUCT.md"

[tool.poetry.dependencies]
python = "^3.8"
django = ">=4.2,<6.0"
django-typer = ">=2.2,<3.0"

rich = { version = ">=10.11.0,<14.0.0", optional = true }
typing-extensions = "^4.12.2"

[tool.poetry.extras]
rich = ["rich"]

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
pytest-django = ">=4.8.0"
pytest-cov = ">=5.0.0"
Sphinx = [
    { version = ">=8.0", markers = "python_version >= '3.10'" },
    { version = ">=7.4", markers = "python_version >= '3.9'" },
    { version = ">=7.0", markers = "python_version < '3.9'" }
]
docutils = [
    { version = ">=0.21", markers = "python_version >= '3.9'" },
    { version = ">=0.20", markers = "python_version < '3.9'" }
]
mypy = ">=1.0"
doc8 = ">=1.1.1"
aiohttp = ">=3.9.1"
readme-renderer = {extras = ["md"], version = ">=42"}
sphinxcontrib-typer = {extras = ["html", "pdf", "png"], version = ">=0.5.1", markers="python_version >= '3.9'"}
django-stubs = ">=4.2.7"
ruff = ">=0.4.1"
furo = ">=2024.7.18"

[tool.mypy]
# The mypy configurations: http://bit.ly/2zEl9WI
allow_redefinition = false
check_untyped_defs = true
disallow_untyped_decorators = false
disallow_any_explicit = false
disallow_any_generics = false
disallow_untyped_calls = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
strict_equality = true
local_partial_types = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
exclude = ["tests"]


# todo doc8 configuration here is not being picked up and doesnt seem to be working
# as expected - for now pass these parameters manually
[tool.doc8]
max-line-length = 100
sphinx = true

[tool.pytest.ini_options]
# py.test options:
DJANGO_SETTINGS_MODULE = "tests.settings.base"
python_files = "test_*.py"
norecursedirs = "*.egg .eggs dist build docs .tox .git __pycache__ shellcompletion"
env = [
    "TERMINAL_WIDTH=80",
]

addopts = [
  "--strict-markers",
  "--cov=django_typer",
  "--cov-branch",
  "--cov-report=term-missing:skip-covered"
]

[tool.coverage.run]
branch = true
source = ["django_typer"]
concurrency = ["multiprocessing"]
parallel = true
relative_files = true
command_line = "-m pytest --cov=django_typer"

[tool.coverage.paths]
source = [
    "django_typer"
]

[tool.pyright]
exclude = ["tests/**/*"]
include = [
    "django_typer"
]

[tool.ruff]
line-length = 88
exclude = [
    "doc",
    "dist",
    "examples"
]

[tool.ruff.lint]
exclude = [
    "tests/**/*"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
